Fixing this is proving very hard
Something in app.test.js is failing. A call to the ws results was definitely not working at some point, but it appears the problem is more complex. Too much to debug. I've tried a lot of fixes to no avail. Going to revert this test coverage for now, but will store this version.